+++ src/main/java/com/github/wintersteve25/tau/components/Tooltip.java
@@ -6,26 +6,30 @@ import com.github.wintersteve25.tau.components.base.PrimitiveUIComponent;
 import com.github.wintersteve25.tau.components.base.UIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
 import com.github.wintersteve25.tau.theme.Theme;
-import com.github.wintersteve25.tau.utils.Color;
 import com.github.wintersteve25.tau.utils.Vector2i;
 import com.mojang.blaze3d.platform.Window;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
-import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.components.Widget;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.network.chat.FormattedText;
-import net.minecraftforge.client.gui.GuiUtils;
+import net.minecraft.world.item.ItemStack;
+import net.minecraftforge.client.ForgeHooksClient;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicReference;
 
 public final class Tooltip implements PrimitiveUIComponent {
 
-    private final List<FormattedText> text;
+    private final List<ClientTooltipComponent> components;
+    private final List<FormattedText> texts;
     private final UIComponent child;
     
-    public Tooltip(List<FormattedText> text, UIComponent child) {
-        this.text = text;
+    public Tooltip(List<ClientTooltipComponent> components, List<FormattedText> texts, UIComponent child) {
+        this.components = components;
+        this.texts = texts;
         this.child = child;
     }
 
@@ -39,10 +43,20 @@ public final class Tooltip implements PrimitiveUIComponent {
         
         Vector2i size = UIBuilder.build(layout, theme, child, renderables, tooltips, dynamicUIComponents, eventListeners);
         Vector2i position = layout.getPosition(size);
+        AtomicReference<List<ClientTooltipComponent>> combined = new AtomicReference<>();
         
         tooltips.add((pPoseStack, pMouseX, pMouseY, pPartialTicks) -> {
             if (Vector2i.within(pMouseX, pMouseY, position.x, position.y, size.x, size.y)) {
-                theme.drawTooltip(pPoseStack, pMouseX, pMouseY, screenWidth, screenHeight, fontRenderer, text);
+                List<ClientTooltipComponent> components = ForgeHooksClient.gatherTooltipComponents(ItemStack.EMPTY, texts, pMouseX, screenWidth, screenHeight, fontRenderer, fontRenderer);
+                
+                if (combined.get() == null)
+                    combined.set(new ArrayList<>(components.size() + this.components.size()));
+                else
+                    combined.get().clear();
+                combined.get().addAll(components);
+                combined.get().addAll(this.components);
+                
+                theme.drawTooltip(pPoseStack, pMouseX, pMouseY, screenWidth, screenHeight, fontRenderer, combined.get());
             } 
         });
 
@@ -50,24 +64,36 @@ public final class Tooltip implements PrimitiveUIComponent {
     }
     
     public static final class Builder {
-        private final List<FormattedText> text;
+        private final List<ClientTooltipComponent> components;
+        private final List<FormattedText> texts;
 
         public Builder() {
-            text = new ArrayList<>();
+            components = new ArrayList<>();
+            texts = new ArrayList<>();
         }
 
         public Builder withText(List<FormattedText> text) {
-            this.text.addAll(text);
+            this.texts.addAll(text);
             return this;
         }
         
         public Builder withText(FormattedText text) {
-            this.text.add(text);
+            this.texts.add(text);
             return this;
         }
-
+        
+        public Builder withComponent(List<ClientTooltipComponent> components) {
+            this.components.addAll(components);
+            return this;
+        }
+        
+        public Builder withComponent(ClientTooltipComponent component) {
+            this.components.add(component);
+            return this;
+        }
+        
         public Tooltip build(UIComponent child) {
-            return new Tooltip(text, child);
+            return new Tooltip(components, texts, child);
         }
     }
 }